Comments, warnings -> ddg exception nodes
Merged in Emery's changes.  Most changes are to comments.  File coyping
code in ddg.procedure is changed.  Warnings now appear in the DDG as
well as in the console.  The function to flush the ddg directory now
only deletes files that match the naming pattern of files that the R
library creates.