Changed .ddg.is.simple to return false for lists
Fixed a bug in .ddg.loadhistory that caused it to fail if the last
line in the file was the timestamp.

Fixed a bug when processing outs.data parameter of ddg.procedure so 
that a variable name can be passed in as a name, not just a string.
Same thing for ddg.data.