Bug fixes, modifications to test cases
Test cases have been changed to deal with bug #40 by removing calls to
substitute. They've also been modified to deal with bug #39 by replacing
calls to Map or mapply with a for loop.

Procedure nodes created through console node now only contain the
abbreviated command as the Value= parameter for ddg.txt. This addresses
bug #37. We might need to eventually remove the Value= parameter for
procedure nodes as it seems useless as of now.

Enhancement #22 has also been addressed. Timestamp data is now stored
within the procedure nodes. DDG Explorer has also been
modified to deal with this change, however the information is not used as
of version 1.011.