Major repairs to ddg.source.
ddg.source was previously ignore the environment under which it was
exectuing. This would lead to a ddg where all values would be null (if
it was sourced with local=T), because the console creation code always
looked in the Global Environment, not the environment of script execution.
This only furtheres my case that we need to seperate the console code from
the source code.

Also repaired  a HUGE error with the timing functions were the working
directory code was not inserted properly.