Fixed error related to lapply; fixed reg expr problem in text xml file
When a function is called via a higher order function, we will see the
ddg associated with the called function.  Not sure that this is
ultimately what we want, but it collapses to a single node and no longer
ends in an error, which it did previously.

Also fixed a bug with regular expressions in the
test-no-instrumentation.xml file.