Temporary commit.
Working on dealing with dev.off as the last line of a function if there
is no explicit opening of a graphics device.  The strategy in this
commit was to annotate code so that ddg.return.value had a call to
ddg.eval as a parameter and passed the statement as a string.  This
causes problems, though, if the statement itself contains return
statements, like inside an if-statement, since when we eval a string
containing return, we end up returning from some function inside RDT
instead of the user's function.

Will try a different strategy, but wanted to hang on to this version
temporarily at least.