Trace output on entry to write functions instead of on exit
When we trace, we are just putting a file name in a list that is then
used to create nodes after the statement completes.  Using on.exit has
problems because if the function itself uses on.exit (like write.csv
does), then that on.exit expression replaces the trace.  Since we are
not actually copying the file at this point, we can do the work on
entry.