Assorted changes for storing values
For connections, store description if no snapshots.

For exceptions, always store the complete message in the node, even if
it is long.

For functions, check the size of the deparsed function (i.e., the text),
rather than the size of the function object when deciding whether to
store a full or partial snapshot.

Changed tests.xml to ignore the hex value associated with an environment
value.