Fixes issue 550
When an inner function does a non-local, non-global assignment, rdtLite
no longer attempts to create a data node for it.  Trying to do so
resulted in an error since the variable could not be found in the global
scope, which is the only scope rdtLite has access to.

Also, rdt creates the correct nodes and edges in this case.