Creates data nodes for variables within environments, like env$var
The node would be labeled as "env$var" either when set or used.  The
code needs to distinguish between use of the $ operator to dereference
environments, as opposed to other uses, as with data frames.  In the
latter case, we want the node created to be labelled with the data frame
name, not the column name.