Added a warning when running an RMarkdown file.
prov.run runs an Rmd file twice: once to collect provenance and once to
create the RMarkdown document.  If the .Rmd file is nondeterministic,
these will not match.

However, if the user runs RMarkdown in RStudio using the interactive
"Run next chunk" commands, the provenance is collected as the script
executes and will match, but then no output document is created.