Updates to work with R 4.0
Some tests have different results.  The default value for
stringsAsFactors changed, so we needed to change the scripts to pass in
the old default.

The plot function was moved into the base package so that changed the
provenance.

Some tests are temporarily disabled because the HF website has expired
certificates, causing the code to fail.