Fixed a bug when tracing input functions.
Some input functions declare a file parameter with no default value.
These functions can often be called by providing a value for a text
parameter and it then reads from the text instead of from a file.  When
we tried to look up the value of the file parameter it was not bound.
We now catch that error, and just don't record any file info in that
case since no file is being read from.  This code is in
.ddg.trace.input.