Tracing shapefiles is in progress
Tracing rgdal readOGR and writeOGR functions is in place.  What is
missing is dealing with the fact that the parameter to these functions
is a folder, not a single file.  We should copy all the files if their
total size is below the max copy limit.

Also, raster files for GIS systems contain multiple files, like
- tif.vat
- auxiliary xml files
- img files have xml files
- tif files are self-contained

Also, if the same file is read multiple times, I am seeing multiple
nodes for the file, although the summarizer only lists it once.  For a
regular text file, we get a single node.  Maybe it is because the file
is not getting copied due to its size???