Changed order of json parts to facilitate testing
When running on GitHub, we should not compare the library list or what
functions came from which libraries.  The way R is installed with GitHub
Actions, it has many more libraries installed then when run from ant.

The order in which functions and libraries are included in the json was
swapped so that it would be easier to write the regular expression
needed to remove the library list.